Feat(item): add bookmark

+++ backend/yam/src/main/java/com/ssafy/yam/domain/bookmark/repository/BookmarkRepository.java
@@ -9,6 +9,6 @@ import org.springframework.data.repository.query.Param;
 public interface BookmarkRepository extends JpaRepository<Bookmark, BookmarkId> {
 
     @Query(value = "select count(*) from bookmark where item_id = :itemId", nativeQuery = true)
-    Long countByItemId(@Param("itemId") Long itemId);
+    int countByItemId(@Param("itemId") int itemId);
 
 }
